Consolidated Python package contents into a single module for simplicity.
